Product group create, update and read

+++ src/main/java/net/vatri/ecommerce/models/ProductGroup.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Table(name = "product_groups")
 public class ProductGroup {
 
-    private Integer id;
+    private Long id;
     private String groupName;
     private String price;
     @Temporal(TemporalType.TIMESTAMP)
@@ -21,11 +20,11 @@ public class ProductGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
